fixed #651
If the user changed the delivery method to takeout and the payment is
card the default tip will be 0%. If the delivery method is delivery and
the payment is card the default tip will be 15% (variable
App.order.tip). If the user had changed the tip value the default value
will be chosed one.