fixed #661
At the javascript side is being used the App.ceil function at the
result of the sum of the products's price (app.js:1281). The same
function is being used to round the values of the tax (app.js:1315) and
tip (app.js:1324). But at the PHP side this function Util::ceil was
called just at the sum of all itens (products price + tax + tip). So I
put the Util::ceil after the sum of the products price (Order.php:45),
after the tax calculation (Order.php:63) and after the tip calculation
(Order.php:58).