Fixed the footer position at the restaurant's page.
The footer position was going up at the restaurant's page. It was happening because the div.restaurant-items has the position:absolute style's attribute and the footer was placed right after the div.body. Now the div.body will have at least the same height that the div.restaurant-items has. (app.js:343)