fixed #669
The orders were set as delivery only if the word "delivery" was used. I changed the logic to be delivery as default and only as takeout if the word "takeout" was used. If neither of them was used, the order will be flagged as for delivery, but also logged in the system, in the Log table to debug further what is wrong with it.