partial #2243 - add the ability to pay using stripe or balanced in the order.php
Updated the Stripe API (php) to the new version (from 1.7.2 to 1.10.1).

Updated the keys (public and secret) of the test env (config.xml). I was afraid to change it from the live one.

I changed the way the user_payment_type was working.
Now the user will be able to have two actives payment types, one for stripe and another for balanced.

When you call the method Crunchbutton_User_Payment_Type::getUserPaymentType() it will return the active payment type for the current processor.

I have also created this method Crunchbutton_User_Payment_Type::processor() returns the current processor type and replaced all the calls from c::config()->processor to this method. That way if we decide to put this data at config’s table we will have to just change this method. Or even is we need to use the processor type that is previously stored that will be easy too.

I created two new files tokenizeCard.stripe.js and tokenizeCard.balanced.js and moved the method App.tokenizeCard to these files