Object-orient metrics internals + slimmer serialization
* splits out the metrics data-producing code into a separate object
  (IMMUTABLE!)
* reduces the number of arguments to each function
* changes serialization format to reflect guarantee that labels are the
  same for every outputted community.
* adds start date and end date to serialization format.

This is in prep for backfilling data and making other changes (e.g.,
warning when start and end date are different than what has been shown
on back end).

It may make more sense to create a large number of classes, rather than
a proliferation of methods on a single huge class. But for now this is
what I'm doing :P.