Update angular chart so that the watches actually work
Options were not updating because the watches were incorrect.

That said, this raises the issue of huge copies when data is large.
We'll have to deal with that in the future (probably by forking
angular-charts and settings $scope.$watch to false and manipulating
object equality ourselves).

NOTE: We switched to use object equality instead (so we can manage memory better)

This appears to have resolved #4874 as well as fix scale changes.